refactor: moved the layout components to their own folders chore: moved the layout components outside of the components folder chore: moved sections to the outside of the components directory test: added tests for the layouts.
