+++ src/components/sections/contact-section.js
@@ -1,6 +1,6 @@
 import React from "react"
-import BackgroundColorLayout from "../layouts/background-color-layout"
-import TextContainerLayout from "../layouts/text-container-layout"
+import BackgroundColorLayout from "../../layouts/background-color-layout/background-color-layout"
+import TextContainerLayout from "../../layouts/text-container-layout/text-container-layout"
 import ExternalLink from "../external-link/external-link"
 
 const ContactSection = () => (
